apps/auth/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pydantic==2.10.3
email-validator==2.2.0
psycopg[binary]==3.2.5
